feat: Update SignUpPageOne component to populate form fields with pageData
The SignUpPageOne component now receives a `pageData` prop, which is used to populate the form fields with the corresponding data. If the `pageData` object contains values for `name`, `email`, `phone`, and `dob`, those values will be used as the initial values for the form fields. Additionally, a check has been added to ensure that the `pageData` object is not empty before setting the form data.

chore: Update SignUpPopover to handle going back to SignUpPageOne

The SignUpPopover component now includes a `handleBack` function that is triggered when the user clicks the back button on SignUpPageTwo. This function sets the `pageOneData` state variable to the current form data, allowing the user to go back to SignUpPageOne and see their previously entered data.